MC2_gesture_detection/requirement.txt
absl-py==0.11.0
appnope==0.1.2
APScheduler==3.6.3
art==5.3
astor==0.8.1
astunparse==1.6.3
backcall==0.2.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2020.11.8
chardet==3.0.4
charset-normalizer==2.0.7
cycler==0.10.0
decorator==5.0.9
flatbuffers==1.12
fonttools==4.25.0
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.10.1
importlib-metadata==2.0.0
ipython==7.27.0
jedi==0.18.0
joblib==1.0.1
Keras==2.4.3
Keras-Applications==1.0.8
keras-flops==0.1.2
Keras-Preprocessing==1.1.2
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282414123/work
Markdown==3.3.3
matplotlib==3.3.3
matplotlib-inline==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
munkres==1.1.4
networkx==2.6.3
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.4.60
opt-einsum==3.3.0
paho-mqtt==1.5.1
pandas==1.3.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.20
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycm==3.3
Pygments==2.10.0
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
python-dateutil==2.8.1
python-telegram-bot==13.7
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
random2==1.0.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.18.3
scikit-learn==0.24.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
telegram==0.0.1
tensorboard==2.4.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-addons==0.12.1
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
termcolor==1.1.0
tflite==2.3.0
threadpoolctl==2.1.0
tifffile==2021.10.12
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
traitlets==5.1.0
typeguard==2.12.0
typing-extensions==3.7.4.3
tzlocal==3.0
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
